# PERF: Series.str.get for pyarrow-backed strings